Don't use effect to draw node in gfx.fx.emerge
This fixes an issue where if the last action leading to victory
was a beta-reduction, the splosion effect would be on top of the
node, instead of in place of the node. This happened because
the victory animation would play concurrently with the emerge
animation, which used an effect to manually draw the node on stage.
However, we now use emerge with nodes that are already on-stage,
so there is no need for this effect.